requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
requests==2.31.0
Werkzeug==2.3.7